Move jpms clearance advice to bootstrap
