+++ dd-java-agent/agent-bootstrap/src/main/java11/datadog/trace/bootstrap/instrumentation/jpms/JpmsAdvisingHelper.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.mule4;
+package datadog.trace.bootstrap.instrumentation.jpms;
 
 import datadog.trace.api.GenericClassValue;
 import java.util.concurrent.atomic.AtomicBoolean;
