+++ dd-java-agent/agent-bootstrap/src/main/java11/datadog/trace/bootstrap/instrumentation/jpms/JpmsClearanceAdvice.java
@@ -1,10 +1,15 @@
-package datadog.trace.instrumentation.mule4;
+package datadog.trace.bootstrap.instrumentation.jpms;
 
-import static datadog.trace.instrumentation.mule4.JpmsAdvisingHelper.ALREADY_PROCESSED_CACHE;
+import static datadog.trace.bootstrap.instrumentation.jpms.JpmsAdvisingHelper.ALREADY_PROCESSED_CACHE;
 
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.implementation.bytecode.assign.Assigner;
 
+/**
+ * This advice can be used by any instrumentation needing that the instrumented class export its
+ * package to the unnamed module (where our things resides). This ensures that an instrumentation
+ * will grant visibility to that class.
+ */
 public class JpmsClearanceAdvice {
   @Advice.OnMethodExit(suppress = Throwable.class)
   public static void openOnReturn(@Advice.This(typing = Assigner.Typing.DYNAMIC) Object self) {
