+++ dd-java-agent/instrumentation/mule-4/build.gradle
@@ -90,9 +90,6 @@ configurations.all {
 }
 
 sourceSets {
-  main_java11 {
-    java.srcDirs "${project.projectDir}/src/main/java11"
-  }
   test {
     output.dir("$buildDir/generated-resources/test", builtBy: 'generateAppResources')
   }
@@ -102,20 +99,6 @@ sourceSets {
   latestDepForkedTest {
     output.dir("$buildDir/generated-resources/latestDepForkedTest", builtBy: 'generateAppResourcesLatest')
   }
-}
-
-compileMain_java11Java.configure {
-  setJavaVersion(it, 11)
-  sourceCompatibility = JavaVersion.VERSION_1_8
-  targetCompatibility = JavaVersion.VERSION_1_8
-}
-
-jar {
-  from sourceSets.main_java11.output
-}
-
-forbiddenApisMain_java11 {
-  failOnMissingClasses = false
 }
 
 tasks.named("compileTestGroovy").configure {
@@ -138,13 +121,7 @@ tasks.named("compileLatestDepForkedTestJava").configure {
 dependencies {
   compileOnly group: 'org.mule.runtime', name: 'mule-core', version: muleVersion
   compileOnly group: 'org.mule.runtime', name: 'mule-tracer-customization-impl', version: muleVersion
-  compileOnly sourceSets.main_java11.output
 
-  main_java11CompileOnly project(':internal-api')
-  main_java11CompileOnly project(':dd-java-agent:agent-tooling')
-  main_java11CompileOnly project(':dd-java-agent:agent-bootstrap')
-
-  testImplementation sourceSets.main_java11.output
   testImplementation project(':dd-java-agent:instrumentation:aws-common')
   testImplementation project(':dd-java-agent:instrumentation:reactor-core-3.1')
   testImplementation project(':dd-java-agent:instrumentation:reactive-streams')
